\name{check.statevector}
\alias{check.statevector}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Check statevector
}
\description{
Check whether or not as given statevector (which are used in some brownie commands) is valid.
}
\usage{
check.statevector(optstr, factvect)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{optstr}{
Character string
}
  \item{factvect}{
Factors whose levels which should exclusively occur in optstr.
}
}
\details{
Mainly an internal function
}
\value{
Logical
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

