\name{clearCommands-methods}
\docType{methods}
\alias{clearCommands}
\alias{clearCommands-methods}
\alias{clearCommands,brownie-method}
\alias{clearCommands,list-method}
\title{ Clear commands}
\description{
Clear all the commands in the commands slot from a brownie object (or list of brownie objects)
}
\section{Methods}{
\describe{

\item{\code{signature(x = "brownie")}}{ Clear from a single brownie object }

\item{\code{signature(x = "list")}}{ Iterate through all brownie objects in this list and clear commands from each }
}}
\keyword{methods}
