\name{deep.phy.copy}
\alias{deep.phy.copy}
\title{
Copy a tree object slot by slot
}
\description{
This is mainly an internal function
}
\usage{
deep.phy.copy(from, to, value, usedata = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{from}{
An S4 object
}
  \item{to}{
An S4 object
}
  \item{value}{
The value to copy
}
  \item{usedata}{
Should data slot be included?
}
}
\details{
This function is not being used currently
}
\value{
A copy of from
}
\author{
J. Conrad Stack
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
