\name{get.nodenames}
\alias{get.nodenames}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read node names from newick text
}
\description{
Internal function to help processing SIMMAP text.
}
\usage{
get.nodenames(newick.txt)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{newick.txt}{
Character string of newick text
}
}
\details{
}
\value{
The node names
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

