\name{getEmptyDataFrame}
\alias{getEmptyDataFrame}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get empty data.frame 
}
\description{
Get an empty data frame with the same structure and column names of the input data frame
}
\usage{
getEmptyDataFrame(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The data frame whose structure is to be copied
}
  \item{...}{
Arguments passed to data.frame() function
}
}
\details{
This is mainly an internal function used to generate subnode dataframes.
}
\value{
An empty data.frame
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

