\name{getSubNodeEdgeInds}
\alias{getSubNodeEdgeInds}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return subnode edge indices
}
\description{
The function gets the node indices on which each subnode is positioned on.
}
\usage{
getSubNodeEdgeInds(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A phylo4d_ext object
}
}
\details{
Mainly an internal function
}
\value{

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
data(parrot)
getSubNodeEdgeInds(parrot[[1]])
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

