\name{hasSubNodes-methods}
\docType{methods}
\alias{hasSubNodes-methods}
\alias{hasSubNodes,phylo-method}
\alias{hasSubNodes,phylo4-method}
\alias{hasSubNodes,phylo4d_ext-method}
\title{Check for subnodes}
\description{
Check if the phylo4d_ext object contains any subnodes
}
\section{Methods}{
\describe{

\item{\code{signature(x = "phylo")}}{ Always returns false }

\item{\code{signature(x = "phylo4")}}{ Always returns false }

\item{\code{signature(x = "phylo4d_ext")}}{ Checks for subnodes }
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s)}
