\name{hasWeight-methods}
\docType{methods}
\alias{hasWeight-methods}
\alias{hasWeight,list-method}
\alias{hasWeight,phylo4d_ext-method}
\title{ Check for tree weights}
\description{
Check if a phylo4d_ext contains tree weights.  It just checks the weight slot
}
\section{Methods}{
\describe{

\item{\code{signature(x = "list")}}{ Check the first item in the list }

\item{\code{signature(x = "phylo4d_ext")}}{ Check for tree weights }
}}
\keyword{methods}
