\name{is.contData}
\alias{is.contData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Is data continuous
}
\description{
Check if the data vector contains continuous data
}
\usage{
is.contData(dat)
}
\arguments{
  \item{dat}{
A data vector
}
}
\details{

}
\value{

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

