\name{nSubNodes}
\alias{nSubNodes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get number of subnodes
}
\description{
Return the number of subnodes in a \code{phylo4d_ext} object
}
\usage{
nSubNodes(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A phylo4d_ext object
}
}
\details{
}
\value{
Return the number of subnodes.
}
\references{
}
\author{
J. Conrad Stack
}
\note{

}

\seealso{
}
\examples{
data(parrot)
nSubNodes(parrot[[1]])
}

\keyword{ methods }

