\name{parrot}
\alias{parrot}
\docType{data}
\title{
Parrotfish data
}
\description{
Dataset of parrotfish jaw morphology (read in from nexus file)
}
\usage{data(parrot)}
\format{
Brownie object
}
\details{
This dataset is from a tutorial titled "Morphological Diversification and Rates of Evolution" by Samantha Price and Peterwainwright presented at 2010 Bodega Applied Phylogenetics Workshop, March 6-13.
}
\source{
\url{http://bodegaphylo.wikispot.org/Morphological_Diversification_and_Rates_of_Evolution}
}
\references{
}
\examples{
data(parrot)
plot(parrot[[1]])
}
\keyword{datasets}
