\name{showBrownie}
\alias{showBrownie}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Show brownie command functions
}
\description{
Shows brownie functions
}
\usage{
showBrownie()
}
%- maybe also 'usage' for other objects documented here.
\details{

}
\value{
Prints text about accessor functions for brownie objects.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

