\name{strip}
\alias{strip}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Strip whitespace 
}
\description{
Internal function:  strip whitespace from the ends of a string
}
\usage{
strip(str, left = TRUE, right = TRUE)
}

\arguments{
  \item{str}{
A string
}
  \item{left}{
Strip white space from left end of string
}
  \item{right}{
Strip white space from right end of string
}
}
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
strip("\tthisis the function\n\n")
}

\keyword{ methods }

