\name{taxaset.names}
\alias{taxaset.names}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get names of the taxasets
}
\description{
Return a character vector with the names of the different taxasets.
}
\usage{
taxaset.names(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A brownie object (or list of brownie objects)
}
}
\details{

}
\value{
A character vector with the taxa set names
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
data(parrot)
taxaset.names(parrot)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

