\name{nSubNodes}
\alias{nSubNodes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get number of subnodes
}
\description{
Return the number of subnodes in a phylo4d_ext object
}
\usage{
nSubNodes(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A phylo4d_ext object
}
}
\details{
}
\value{
The number of subnodes
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
J. Conrad Stack
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
data(parrot)
nSubNodes(parrot[[1]])
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

